Fixes for signed/unsigned conversion

+++ examples/c/mechanic_module_ex_poolsetup.c
@@ -107,7 +107,7 @@ int PoolProcess(pool **all, pool *current, void *s) {
 
   Message(MESSAGE_COMMENT, "Pool %d processed with option ilimit %d\n", current->pid, ilimit);
 
-  if (current->pid < max_pools) return POOL_CREATE_NEW;
+  if (current->pid < (unsigned int)max_pools) return POOL_CREATE_NEW;
   return POOL_FINALIZE;
 }
